test(domain-routing): split routing context suites
Constraint: Preserve all 23 routing-context scenarios and assertions in test-only modules.
Rejected: Duplicate shared fixtures across modules | keeping fixture construction centralized avoids drift.
Confidence: high
Scope-risk: narrow
Directive: Keep contract, phrase-matching, and resolver scenarios in their cohesive modules.
Tested: 23 owned tests; 145 domain discovery tests; compileall; Ruff; git diff --check.
Not-tested: Full repository test discovery outside the domain test surface.
